pyproject.toml
[project]
name = "llmcord"
version = "0.1.0"
description = "A Discord bot for LLM interaction"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp",
    "httpx",
    "loguru",
    "pyyaml",
    "pydantic>=2.0.0",
    "openai",
    "types-PyYAML",
    "ddgs",
    "litellm",
    "trafilatura",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 320
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["llmcord"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["llmcord"]
typeCheckingMode = "strict"
pythonVersion = "3.13"
reportMissingTypeStubs = true
reportUnknownMemberType = true
reportUnknownVariableType = true
reportUnknownArgumentType = true
reportUnknownParameterType = true
reportUnknownReturnType = true
reportUnnecessaryTypeIgnoreComment = true
reportUnusedImport = true
reportUnusedVariable = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportImplicitStringConcatenation = true